More refactoring
- Refactored the remaining drawing functions from JBlockDiagramVisualizer to  ModifiedElkNodePainter

- Deleted JElkGraph & refactored its usage

TODO: Check for bugs & clean up formatting
