+++ src/main/scala/edg_ide/swing/JBlockDiagramVisualizer.scala
@@ -1,120 +1,125 @@
 package edg_ide.swing
 
-import com.intellij.util.ui.UIUtil
-import edg.wir.DesignPath
-import edg_ide.edgir_graph.ElkEdgirGraphUtils
-import org.eclipse.elk.graph.{ElkEdge, ElkGraphElement, ElkNode, ElkPort}
+import org.eclipse.elk.graph.{ElkGraphElement, ElkNode}
 
-import java.awt.geom.Rectangle2D
-import java.awt.image.BufferedImage
-import java.awt.{BasicStroke, Color, Graphics2D, TexturePaint}
-import scala.jdk.CollectionConverters.CollectionHasAsScala
+import java.awt.event.{MouseAdapter, MouseEvent}
+import java.awt.{Dimension, Graphics, Rectangle}
+import javax.swing.{JComponent, Scrollable}
+import scala.collection.mutable
 
 
-/** Block diagram visualizer that customizes the rendering in JElkGraph with options specific to
-  * design block diagrams:
-  * - tunnel link names by heuristic matching (degenerate self-to-self links)
-  * - additional setError(elts) to render elts as filled in red
-  * - additional setStable(elts) to render elts as stale (???)
-  */
-class JBlockDiagramVisualizer(rootNode: ElkNode, showTop: Boolean = false) extends
-    JElkGraph(rootNode, showTop) {
-  protected val hatchRect = new Rectangle2D.Double(0, 0, 16, 16)
-  def makeHatchTexture(backgroundColor: Color, foregroundColor: Color): TexturePaint = {
-    val hatchImage = new BufferedImage(hatchRect.width.toInt, hatchRect.height.toInt, BufferedImage.TYPE_INT_ARGB)
-    val hatchGraphics = hatchImage.createGraphics()
-    val hatchTexture = new TexturePaint(hatchImage, hatchRect)
-    hatchGraphics.setColor(backgroundColor)
-    hatchGraphics.fill(hatchRect)
-    hatchGraphics.setColor(foregroundColor)
-    hatchGraphics.setStroke(new BasicStroke(2))
-    hatchGraphics.drawLine(0, 16, 16, 0)
-    hatchTexture
+/** Block diagram visualizer that customizes the rendering with options specific to
+ * design block diagrams:
+ * - tunnel link names by heuristic matching (degenerate self-to-self links)
+ * - additional setError(elts) to render elts as filled in red
+ * - additional setStable(elts) to render elts as stale (???)
+ */
+class JBlockDiagramVisualizer(var rootNode: ElkNode, var showTop: Boolean = false)
+  extends JComponent with Scrollable with Zoomable {
+  private val elementToolTips = mutable.Map[ElkGraphElement, String]()
+  private val margin: Int = 32 // margin in pixels, regardless of zoom level, so tunnel labels aren't cut off
+  protected var highlighted: Option[Set[ElkGraphElement]] = None
+  protected var errorElts: Set[ElkGraphElement] = Set()
+  protected var staleElts: Set[ElkGraphElement] = Set()
+  private var zoomLevel: Float = 1.0f
+  private var painter = new ModifiedElkNodePainter(rootNode, showTop)
+
+  override def getZoom = zoomLevel
+
+  override def setZoom(zoom: Float): Unit = {
+    zoomLevel = zoom
   }
 
+  def setSelected(elts: Set[ElkGraphElement]): Unit = {
+    painter.setSelected(elts)
+    validate()
+    repaint()
+  }
 
-  protected var errorElts: Set[ElkGraphElement] = Set()
+  def setHighlighted(elts: Option[Set[ElkGraphElement]]): Unit = {
+    highlighted = elts
+    painter.setHighlighted(highlighted)
+    validate()
+    repaint()
+  }
+
+  setGraph(rootNode)
+
+  def getGraph: ElkNode = rootNode
+
+  def setGraph(newGraph: ElkNode): Unit = {
+    errorElts = Set()
+    staleElts = Set()
+    painter = new ModifiedElkNodePainter(newGraph, showTop, zoomLevel)
+    elementToolTips.clear()
+    highlighted = None
+    painter.setSelected(Set())
+    rootNode = newGraph
+    revalidate()
+    repaint()
+  }
+
   def setError(elts: Set[ElkGraphElement]): Unit = {
     errorElts = elts
     validate()
     repaint()
   }
 
-  protected var staleElts: Set[ElkGraphElement] = Set()
   def setStale(elts: Set[ElkGraphElement]): Unit = {
     staleElts = elts
     validate()
     repaint()
   }
 
-  override def setGraph(newGraph: ElkNode): Unit = {
-    errorElts = Set()
-    staleElts = Set()
-    super.setGraph(newGraph)
+  def getElementForLocation(x: Int, y: Int): Option[ElkGraphElement] = {
+    painter.getElementForLocation(x, y)
   }
 
-  override def fillGraphics(base: Graphics2D, background: Color, element: ElkGraphElement): Graphics2D = {
-    if (errorElts.contains(element)) {
-      val newBase = base.create().asInstanceOf[Graphics2D]
-      newBase.setColor(blendColor(background, Color.RED, 0.25))
-      newBase  // explicitly ignores showTop invisibility if it's an error
-    } else if (staleElts.contains(element)) {
-      val newBase = base.create().asInstanceOf[Graphics2D]
-      if (highlighted.isDefined && !highlighted.get.contains(element)) { // dimmed out if not highlighted
-        newBase.setPaint(makeHatchTexture(background, blendColor(background, base.getColor, 0.0375)))
-      } else {
-        newBase.setPaint(makeHatchTexture(background, blendColor(background, base.getColor, 0.15)))
+  override def paintComponent(paintGraphics: Graphics): Unit = {
+    painter = new ModifiedElkNodePainter(rootNode, showTop, zoomLevel, errorElts, staleElts)
+    painter.paintComponent(paintGraphics, this.getBackground)
+  }
+
+  // support for mouse drag: https://docs.oracle.com/javase/tutorial/uiswing/components/scrollpane.html
+  setAutoscrolls(true)
+  // TODO proper drag support
+
+
+  addMouseListener(new MouseAdapter {
+    override def mousePressed(e: MouseEvent): Unit = {
+      requestFocusInWindow()
+    }
+  })
+
+  // Tooltip operations
+  //
+  def setElementToolTip(element: ElkGraphElement, text: String): Unit = {
+    elementToolTips.put(element, text)
+  }
+
+  override def getToolTipText(e: MouseEvent): String = {
+    painter.getElementForLocation(e.getX, e.getY) match {
+      case None => null
+      case Some(element) => elementToolTips.get(element) match {
+        case None => null
+        case Some(text) => text
       }
-      newBase
-    } else {
-      super.fillGraphics(base, background, element)
     }
   }
 
+  override def getPreferredScrollableViewportSize: Dimension = getPreferredSize
 
+  // Scrollable APIs
+  //
+  override def getPreferredSize: Dimension =
+    new Dimension((rootNode.getWidth * zoomLevel + 2 * margin).toInt,
+      (rootNode.getHeight * zoomLevel + 2 * margin).toInt)
 
-  override def paintEdge(parentG: Graphics2D, blockG: Graphics2D, background: Color, edge: ElkEdge): Unit = {
-    super.paintEdge(parentG, blockG, background, edge)
+  override def getScrollableBlockIncrement(rectangle: Rectangle, i: Int, i1: Int): Int = 1
 
-    // TODO dedup from JElkGraph
-    val thisG = if (edge.getSources == edge.getTargets) {
-      val edgeTargetBlockOption = edge.getSources.asScala.headOption.collect {
-        case sourcePort: ElkPort => sourcePort.getParent
-      }
-      if (edgeTargetBlockOption == Some(edge.getContainingNode)) {
-        parentG
-      } else {
-        blockG
-      }
-    } else {
-      blockG
-    }
+  override def getScrollableUnitIncrement(rectangle: Rectangle, i: Int, i1: Int): Int = 1
 
-    if (edge.getSources == edge.getTargets) {  // degenerate, "tunnel" (by heuristic / transform) edge
-      val label = edge.getProperty(ElkEdgirGraphUtils.DesignPathMapper.property) match {
-        case DesignPath(steps) => steps.lastOption.getOrElse("")
-        case _ => ""
-      }
+  override def getScrollableTracksViewportWidth: Boolean = false
 
-      val targetPointOpt = edge.getSections.asScala.headOption.map { section =>
-        val bend = section.getBendPoints.asScala.head
-        (bend.getX, bend.getY, section.getStartX, section.getStartY)
-      }
-
-      val textG = textGraphics(thisG, background, edge)
-      targetPointOpt match {
-        case Some((x, y, x1, y1)) if (x1 == x) && (y > y1) =>
-          DrawAnchored.drawLabel(textG, label, (x, y), DrawAnchored.Top)
-        case Some((x, y, x1, y1)) if (x1 == x) && (y < y1) =>
-          DrawAnchored.drawLabel(textG, label, (x, y), DrawAnchored.Bottom)
-        case Some((x, y, x1, y1)) if (y1 == y) && (x > x1) =>
-          DrawAnchored.drawLabel(textG, label, (x, y), DrawAnchored.Left)
-        case Some((x, y, x1, y1)) if (y1 == y) && (x < x1) =>
-          DrawAnchored.drawLabel(textG, label, (x, y), DrawAnchored.Right)
-        case Some((x, y, _, _)) =>
-          DrawAnchored.drawLabel(textG, label, (x, y), DrawAnchored.Center)
-        case None =>
-      }
-    }
-  }
+  override def getScrollableTracksViewportHeight: Boolean = false
 }
